+++ src/main/java/nextstep/security/authorization/manager/DenyAllAuthorizationManager.java
@@ -1,0 +1,12 @@
+package nextstep.security.authorization.manager;
+
+import nextstep.security.authentication.Authentication;
+import nextstep.security.authorization.AuthorizationDecision;
+
+public class DenyAllAuthorizationManager<T> implements AuthorizationManager<T> {
+
+    @Override
+    public AuthorizationDecision check(Authentication authentication, T object) {
+        return AuthorizationDecision.ACCESS_DENIED;
+    }
+}
